// src/components/PanelComponents/FormEditor/types.ts
export type TUploadDefaultType = Array<{
  uid: string;
  name: string;
  status: string;
  url: string;
}>;
// ...
export interface IDataListConfigType {
  key: string;
  name: string;
  type: 'DataList';
  cropRate: number;
}

export type TDataListDefaultTypeItem = {
  id: string;
  title: string;
  desc: string;
  link: string;
  type?: number;
  imgUrl: TUploadDefaultType;
};

export type TDataListDefaultType = Array<TDataListDefaultTypeItem>;
// ...
export interface IRadioConfigType<KeyType> {
  key: string;
  name: string;
  type: 'Radio';
  range: Array<{
    key: KeyType;
    text: string;
  }>;
}
export type TRadioDefaultType<KeyType> = KeyType;

///////////////

export interface ISwitchConfigType {
  key: string;
  name: string;
  type: 'Switch';
}
export type TSwitchDefaultType = boolean;